requirements.txt
fastapi
python-multipart
jinja2
uvicorn
pymongo
pyjwt
pytelegrambotapi
python-dateutil